Refactor: split monolithic file into multiple smaller files
- Separated bootstrap connection and server
- Moved bootstrapServer to bootstrap_node.go
- Moved connectBootstrapNode to bootstrap_connection.go
